Move access restricted test to its own file
