+++ corehq/messaging/scheduling/scheduling_partitioned/tests/test_dbaccessors_partitioned.py
@@ -138,45 +138,6 @@ class TestSchedulingPartitionedDBAccessorsGetAndSave(BaseSchedulingPartitionedDB
         with self.assertRaises(TimedScheduleInstance.DoesNotExist):
             get_timed_schedule_instance(uuid.uuid4())
 
-    def test_access_restricted(self):
-        with self.assertRaises(AccessRestricted):
-            self.make_alert_schedule_instance(uuid.uuid4()).save()
-
-        instance = self.make_alert_schedule_instance(uuid.uuid4())
-        save_alert_schedule_instance(instance)
-        with self.assertRaises(AccessRestricted):
-            instance.delete()
-
-        with self.assertRaises(AccessRestricted):
-            [obj for obj in AlertScheduleInstance.objects.all()]
-
-        with self.assertRaises(AccessRestricted):
-            AlertScheduleInstance.objects.all()[0]
-
-        with self.assertRaises(AccessRestricted):
-            AlertScheduleInstance.objects.all()[0:10]
-
-        with self.assertRaises(AccessRestricted):
-            len(AlertScheduleInstance.objects.all())
-
-        with self.assertRaises(AccessRestricted):
-            AlertScheduleInstance.objects.count()
-
-        with self.assertRaises(AccessRestricted):
-            AlertScheduleInstance.objects.filter(schedule_instance_id=None)
-
-        with self.assertRaises(AccessRestricted):
-            AlertScheduleInstance.objects.exclude(schedule_instance_id=None)
-
-        with self.assertRaises(AccessRestricted):
-            AlertScheduleInstance.objects.get(schedule_instance_id=None)
-
-        with self.assertRaises(AccessRestricted):
-            AlertScheduleInstance.objects.create(schedule_instance_id=None)
-
-        with self.assertRaises(AccessRestricted):
-            AlertScheduleInstance.objects.get_or_create(schedule_instance_id=None)
-
 
 class TestSchedulingPartitionedDBAccessorsDeleteAndFilter(BaseSchedulingPartitionedDBAccessorsTest):
